+++ src/main/java/com/danang_auction/controller/AuctionController.java
@@ -1,7 +1,6 @@
 package com.danang_auction.controller;
-
-import com.danang_auction.service.AssetService;
 import com.danang_auction.exception.ResourceNotFoundException;
+import com.danang_auction.service.AssetService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -17,9 +16,9 @@ public class AuctionController {
     private final AssetService assetService;
 
     @PutMapping("/{id}/approve")
-    public ResponseEntity<String> approveAsset(@PathVariable Integer id) {
+    public ResponseEntity<String> approveAsset(@PathVariable Long id) {
         try {
-            assetService.approveAsset(id);
+            assetService.approveAsset(id); // Đã đổi sang Long
             return ResponseEntity.ok("Tài sản đã được phê duyệt");
         } catch (ResourceNotFoundException e) {
             return ResponseEntity.status(404).body(e.getMessage());
